+++ src/auth/auth.controllers.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import { userLogin, userRegister } from "../auth/auh.service";
+import { userLogin } from "../auth/auh.service";
 
 export const Login = async (req: Request, res: Response) => {
   const { email, password } = req.body;
@@ -15,19 +15,5 @@ export const Login = async (req: Request, res: Response) => {
     return res.json(loginResult);
   } catch (error) {
     return res.status(401).json({ message: "Error al loguearse" });
-  }
-};
-
-export const Register = async (req: Request, res: Response) => {
-  const { firstname, lastname, email, password } = req.body;
-
-  try {
-    const newUser = await userRegister(firstname, lastname, email, password);
-    res
-      .status(200)
-      .json({ message: "Usuario creado con éxito", user: newUser });
-  } catch (error) {
-    console.error("Error al crear el usuario:", error);
-    res.status(500).json({ error: "Error al crear el usuario" });
   }
 };
