+++ src/client/client.controllers.ts
@@ -1,0 +1,15 @@
+import { Request, Response } from "express";
+import { clientRegister } from "../client/client.service";
+
+export const Client = async (req: Request, res: Response) => {
+  const { firstname, lastname, email, password } = req.body;
+  try {
+    const newUser = await clientRegister(firstname, lastname, email, password);
+    res
+      .status(200)
+      .json({ message: "Usuario creado con éxito", user: newUser });
+  } catch (error) {
+    console.error("Error al crear el usuario:", error);
+    res.status(500).json({ error: "Error al crear el usuario" });
+  }
+};
