+++ src/client/client.routes.ts
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import { Client } from "../client/client.controllers";
+
+const clientRoutes = Router();
+
+clientRoutes.post("/register", Client);
+
+export default clientRoutes;
