+++ src/routes/index.routes.ts
@@ -1,8 +1,14 @@
 import { Router } from "express";
 import authRoutes from "../auth/auth.routes";
+import userRoutes from "../user/user.routes";
+import clientRoutes from "../client/client.routes";
 
 const indexRoutes = Router();
 
 indexRoutes.use("/auth", authRoutes);
+
+indexRoutes.use("/user", userRoutes);
+
+indexRoutes.use("/client", clientRoutes);
 
 export default indexRoutes;
