+++ src/user/user.routes.ts
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import { Register } from "./user.controllers";
+
+const userRoutes = Router();
+
+userRoutes.post("/register", Register);
+
+export default userRoutes;
